gsl: Provide CMake support file resource
Add GSLCMakeSupport repo as a resource to the project. This supplies
CMake "module" support files for find_package.

Stage, build and install resource as part of install of GSL.

Add a test block to create a mini-project that uses these files to
configure build and run a simple GSL program.