qt5base: Add svg submodule as a resource
Bayeux's Qt gui uses QtSVG, so vendor it as a resource.

NB: Qt's submodules like (if not require) to be installed alongside
each other. To balance modularity against ease of use, add the SVG
submodule as a resource rather than a separate formula. Similarly,
don't use main qt5 formula because that is very heavyweight for what
is currently required.