Fix build errors on Ubuntu 16.04
Not 100% diagnosed, but appears to be due to Qt5's configure command gets mixed up
between system and brewed ICU libraries, and misses out needed `-rpath` and similar
flags.

Make dependency on icu4c explicit, and add needed `-R` flags to QT5 configure on Linux.
Tested on Ubuntu16.04, needs further checks on other platforms. It may also be related to the
use of the gold linker on 16.