falaise-particleid: new formulae
Preliminary Formula for ParticleIdentification module under development
for Falaise's flreconstruct application

- https://github.com/xgarrido/ParticleIdentification

The formula is 'HEAD' only as the is no stable tag yet, though the
module is ready to try out. A patch is supplied to get it building
under brew/Falaise 3.

The Falaise formula is updated to create the `lib/Falaise/modules`
directory as a full rather than softlinked path into the falaise
formula. This is a test of allowing modules to be installed
independently of Falaise. Though functional for installing Falaise
and ParticleIdentification, it cannot run the PID module without
pointing the flreconstruct script to it. This is due to flreconstruct
only considering the directory under its Cellar location. This
requires further review.