Import GSL Formula, bumping revisions of users
Changes to GSL soversion by upstream caused issues with users of
the package, as reported in #22.

Import GSL formula @2.4 into this tap to provide stability as it is
a major dependency of SuperNEMO software. Bump revision numbers
of using packages to ensure these will be reinstalled on upgrade.

Update supplied formulae list in README.